feat(AnswerPage):  - refacts the QuestionDetail page transforming it at AnswerPage with better use of the componentizing  - creates a StatusMessageProvider to handle general messages of the app  - creates a AnswersMessageProvider to handle shared answers' state between Answer page's components  - creates a StakingProvider to provide some staking states through the Answer page's components
