+++ frontend/react/src/pages/AnswerPage/types.ts
@@ -1,0 +1,22 @@
+export type Question = {
+  id: string;
+  title: string;
+  content: string;
+  authorAddress: string;
+  authorName: string;
+  timestamp: string;
+  stakeAmount: string;
+  tags: string[];
+  repositoryUrl: string;
+  isOpen: boolean;
+}
+
+export type Answer = {
+  id: string
+  authorAddress: string
+  authorName: string
+  content: string
+  timestamp: string
+  isCorrect: boolean
+  votes: number
+}
