+++ frontend/react/src/providers/StatusMessageProvider/statusMessageContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+import { StatusMessage } from "./types";
+
+interface StatusMessageContextProps {
+  statusMessage: StatusMessage | null
+  setStatusMessage: (statusMessage: StatusMessage | null) => void
+}
+
+export const StatusMessageContext = createContext({} as StatusMessageContextProps)
