+++ frontend/react/vite.config.ts
@@ -9,6 +9,8 @@ export default defineConfig({
     alias: {
       "@logos": fileURLToPath(new URL("src/assets/logos", import.meta.url)),
       "@components": fileURLToPath(new URL("src/components", import.meta.url)),
+      "@providers": fileURLToPath(new URL("src/providers", import.meta.url)),
+      "@utils": fileURLToPath(new URL("src/utils", import.meta.url)),
     },
   },
 });
